+++ src/main/java/com/huysor/ecommerce/phoneshop/exception/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package com.huysor.ecommerce.phoneshop.exception;
 import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiException {
-	public ResourceNotFoundException(String resource, Integer id) {
+	public ResourceNotFoundException(String resource, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s with id = %d not found", resource, id));
 	}
 }
