+++ src/main/java/com/huysor/ecommerce/phoneshop/repository/ModelRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ModelRepository extends JpaRepository<Model, Integer> {
-List<Model>findByBrandsId(Integer brandsId);
+public interface ModelRepository extends JpaRepository<Model,Long> {
+List<Model>findByBrandsId(Long brandsId);
 }
